refactor(organisms): merge SfAccordion
