--- packages/vue/src/components/organisms/SfAccordion/SfAccordion.html
@@ -1,4 +1,0 @@
-<div class="sf-accordion" :class="{'sf-accordion--has-chevron': showChevron}">
-  <!--@slot default slot to setup SfAccordionItem elements -->
-  <slot />
-</div>
